YARN-5925. Extract hbase-backend-exclusive utility methods from TimelineStorageUtil. Contributed by Haibo Chen.
(cherry picked from commit 8288030cb4aa3b5a9425cc0a3f6df03a3eae1dfb)
